refactor(user): store money as decimal yuan instead of int cents
Money columns move to decimal(12,2); Go entities become float64 while
output DTOs keep emitting %.2f strings, so API response shapes are
unchanged. The cents apparatus goes away: ParseDecimalCents and
MulInt32 are removed (AddInt32 stays for other callers), the
MoneyCents marker field is dropped, admin balance adjustments parse
yuan directly (max two decimals, negatives allowed), and all five
formatting sites drop their /100 conversions. Score columns are
untouched.

+++ app/common/model/user.go
@@ -14,7 +14,7 @@ type User struct {
 	Avatar        string    `gorm:"column:avatar;not null;comment:头像" json:"avatar"`                                                      // 头像
 	Gender        int32     `gorm:"column:gender;not null;comment:性别:0=未知,1=男,2=女" json:"gender"`                                         // 性别:0=未知,1=男,2=女
 	Birthday      time.Time `gorm:"column:birthday;comment:生日" json:"birthday"`                                                           // 生日
-	Money         int32     `gorm:"column:money;not null;comment:余额" json:"money"`                                                        // 余额
+	Money         float64   `gorm:"column:money;not null;comment:余额" json:"money"`                                                        // 余额
 	Score         int32     `gorm:"column:score;not null;comment:积分" json:"score"`                                                        // 积分
 	LastLoginTime int64     `gorm:"column:last_login_time;comment:上次登录时间" json:"last_login_time"`                                         // 上次登录时间
 	LastLoginIP   string    `gorm:"column:last_login_ip;not null;comment:上次登录IP" json:"last_login_ip"`                                    // 上次登录IP
